backend/app/requirements.txt
wheel==0.37.1
asgiref==3.5.2
Django==4.1.2
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.1
pytz==2022.1
sqlparse==0.4.3
typing-extensions==4.3.0
django-jazzmin==2.6.0
dill==0.3.6
astroid==2.12.13
djangorestframework==3.14.0
isort==5.11.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.6.2
tomli==2.0.1
tomlkit==0.11.6
wrapt==1.14.1